clean up duplicated kube test env code
