+++ pkg/ucp/queue/apiserver/client_test.go
@@ -6,12 +6,9 @@
 package apiserver
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"fmt"
-	"os"
-	"os/exec"
 	"path/filepath"
 	"testing"
 	"time"
@@ -19,14 +16,11 @@ import (
 	"github.com/project-radius/radius/pkg/ucp/queue/client"
 	v1alpha1 "github.com/project-radius/radius/pkg/ucp/store/apiserverstore/api/ucp.dev/v1alpha1"
 	"github.com/project-radius/radius/pkg/ucp/util/testcontext"
+	"github.com/project-radius/radius/test/ucp/kubeenv"
 	"github.com/stretchr/testify/require"
-	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
-	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
 	runtimeclient "sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/envtest"
 )
 
 const pollingInterval = time.Duration(100) * time.Millisecond
@@ -129,7 +123,8 @@ func TestGenerateID(t *testing.T) {
 }
 
 func TestClient(t *testing.T) {
-	rc, env, err := startEnvironment()
+	rc, env, err := kubeenv.StartEnvironment([]string{filepath.Join("..", "..", "..", "..", "deploy", "Chart", "crds", "ucpd")})
+
 	require.NoError(t, err, "If this step is failing for you, run `make test` inside the repository and try again. If you are still stuck then ask for help.")
 	defer func() {
 		_ = env.Stop()
@@ -139,7 +134,7 @@ func TestClient(t *testing.T) {
 	defer cancel()
 
 	ns := "radius-test"
-	err = ensureNamespace(ctx, rc, ns)
+	err = kubeenv.EnsureNamespace(ctx, rc, ns)
 	require.NoError(t, err)
 
 	testLockTime := time.Duration(1) * time.Second
@@ -311,65 +306,4 @@ func TestClient(t *testing.T) {
 
 		require.Equal(t, msgCount, recvCnt)
 	})
-}
-
-func startEnvironment() (runtimeclient.Client, *envtest.Environment, error) {
-	assetDir, err := getKubeAssetsDir()
-	if err != nil {
-		return nil, nil, err
-	}
-
-	testEnv := &envtest.Environment{
-		CRDDirectoryPaths:     []string{filepath.Join("..", "..", "..", "..", "deploy", "Chart", "crds", "ucpd")},
-		ErrorIfCRDPathMissing: true,
-		BinaryAssetsDirectory: assetDir,
-	}
-
-	scheme := runtime.NewScheme()
-
-	utilruntime.Must(clientgoscheme.AddToScheme(scheme))
-	utilruntime.Must(v1alpha1.AddToScheme(scheme))
-
-	cfg, err := testEnv.Start()
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to initialize environment: %w", err)
-	}
-
-	client, err := runtimeclient.New(cfg, runtimeclient.Options{
-		Scheme: scheme,
-	})
-	if err != nil {
-		_ = testEnv.Stop()
-		return nil, nil, fmt.Errorf("failed to create kubernetes client: %w", err)
-	}
-
-	return client, testEnv, nil
-}
-
-func getKubeAssetsDir() (string, error) {
-	assetsDirectory := os.Getenv("KUBEBUILDER_ASSETS")
-	if assetsDirectory != "" {
-		return assetsDirectory, nil
-	}
-
-	// We require one or more versions of the test assets to be installed already. This
-	// will use whatever's latest of the installed versions.
-	cmd := exec.Command("setup-envtest", "use", "-i", "-p", "path", "--arch", "amd64")
-	var out bytes.Buffer
-	cmd.Stdout = &out
-	err := cmd.Run()
-	if err != nil {
-		return "", fmt.Errorf("failed to call setup-envtest to find path: %w", err)
-	} else {
-		return out.String(), err
-	}
-}
-
-func ensureNamespace(ctx context.Context, client runtimeclient.Client, namespace string) error {
-	nsObject := v1.Namespace{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: namespace,
-		},
-	}
-	return client.Create(ctx, &nsObject, &runtimeclient.CreateOptions{})
 }
